Made rf_attr_sel more robust for Py3
Cast the input to np.float32 to avoid problems with RF

+++ py/rf_attr_selection.py
@@ -74,7 +74,7 @@ if __name__ == '__main__':
     if sys.version_info[0] == 2:
         data = np.genfromtxt(sys.stdin, delimiter=",", dtype=np.float32)
     elif sys.version_info[0] == 3:
-        data= np.array([list(map(float,l.split(","))) for l in sys.stdin])
+        data= np.array([list(map(float,l.split(","))) for l in sys.stdin], dtype=np.float32)
     else:
         raise Exception('Unrecognized version of Python: %s' % str(sys.version_info))
